Mv ContentSpec to models.models
