+++ ansible_galaxy/installed_repository_db.py
@@ -4,7 +4,7 @@ import os
 from ansible_galaxy import matchers
 from ansible_galaxy import installed_namespaces_db
 from ansible_galaxy.models.content_repository import ContentRepository
-from ansible_galaxy.models.content_spec import ContentSpec
+from ansible_galaxy.models.models import ContentSpec
 
 log = logging.getLogger(__name__)
