+++ tests/ansible_galaxy/models/test_content_repository_model.py
@@ -6,21 +6,21 @@ import pytest
 import six
 
 from ansible_galaxy.models import content_repository
-from ansible_galaxy.models import content_spec
+from ansible_galaxy.models.models import ContentSpec
 
 log = logging.getLogger(__name__)
 
 
 @pytest.fixture
 def csf(request):
-    _cs = content_spec.ContentSpec(namespace='somenamespace', name='somename', version='1.2.3')
+    _cs = ContentSpec(namespace='somenamespace', name='somename', version='1.2.3')
     yield _cs
 
 
 def test_frozen(csf):
     cr = content_repository.ContentRepository(content_spec=csf)
 
-    new_cr = content_spec.ContentSpec(namespace='somenamespace', name='somename', version='1.2.3')
+    new_cr = ContentSpec(namespace='somenamespace', name='somename', version='1.2.3')
 
     with pytest.raises(attr.exceptions.FrozenInstanceError):
         cr.content_spec = new_cr
@@ -70,11 +70,11 @@ def CR(content_spec=None, path=None):
     return content_repository.ContentRepository(content_spec=content_spec)
 
 
-ns_n = content_spec.ContentSpec(namespace='ns', name='n', version='1.0.0')
-diffns_n = content_spec.ContentSpec(namespace='diffns', name='n', version='1.0.0')
+ns_n = ContentSpec(namespace='ns', name='n', version='1.0.0')
+diffns_n = ContentSpec(namespace='diffns', name='n', version='1.0.0')
 ns_n_1_0_0 = ns_n
 diffns_n_1_0_0 = diffns_n
-ns_n_1_0_1 = content_spec.ContentSpec(namespace='ns', name='n', version='1.0.1')
+ns_n_1_0_1 = ContentSpec(namespace='ns', name='n', version='1.0.1')
 
 
 path1 = '/dev/null/1'
